feat: restructure folders
