+++ src/testing-utils/index.ts
@@ -1,0 +1,2 @@
+export * from './contract-utils';
+export * from './testing-utils';
